# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marasem API",
        "description": "This API supports social media login with Google, Facebook, and Behance. Users are redirected to the provider's login page and upon successful authentication, they are logged into the app and issued a token.",
        "contact": {
            "email": "amrfayez.247@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server for API routes"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server for Web routes"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/collections-tags": {
            "get": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Get collections and tags with their usage count",
                "operationId": "a8fec9535fa7beb6555d12f92932569d",
                "responses": {
                    "200": {
                        "description": "Collections and tags fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collections": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Collection"
                                            }
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tag"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/artworks/{id}/like": {
            "post": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Like an artwork",
                "operationId": "9dac56176d45662d8cd950437949f633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the artwork to like",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Artwork liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Liked successfully"
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artwork not found"
                    },
                    "400": {
                        "description": "Already liked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artworks"
                ],
                "summary": "Unlike an artwork",
                "operationId": "579ce77e802dcd12914013b29dc9fd3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the artwork to unlike",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artwork unliked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unliked successfully"
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artwork or like not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "required": [
                    "id",
                    "name",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the tag",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the tag",
                        "type": "string",
                        "example": "Landscape"
                    },
                    "category_id": {
                        "description": "The ID of the category the tag belongs to",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "Collection": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the collection",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the collection",
                        "type": "string",
                        "example": "Nature Collection"
                    },
                    "tags": {
                        "description": "An array of tags associated with the collection",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Landscape"
                        },
                        "nullable": true
                    },
                    "followers": {
                        "description": "The number of followers of the collection",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your Bearer token to access secured endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}